refactor tests into two types
