--- services/workflow_chat/tests_to_run.txt
@@ -1,10 +1,0 @@
-test_inputs/empty.yaml test_inputs/history_gen_yaml_water_bug.json
-test_inputs/empty.yaml test_inputs/history_gen_yaml_simple_lang_bug.json
-test_inputs/empty.yaml test_inputs/history_gen_yaml_input.json
-test_inputs/empty.yaml test_inputs/history_gen_yaml_trigger_nodes.json
-test_inputs/existing_yaml_gen_yaml_convo_turn.yaml test_inputs/history_gen_yaml_convo_turn.json
-test_inputs/existing_yaml_gen_yaml_edit_job_code.yaml test_inputs/history_gen_yaml_edit_job_code.json 
-test_inputs/existing_yaml_gen_yaml_input_second_turn.yaml test_inputs/history_gen_yaml_input_second_turn.json
-test_inputs/existing_yaml_gen_yaml_ask_job_code.yaml test_inputs/history_gen_yaml_ask_job_code.json
-test_inputs/existing_yaml_gen_yaml_error.yaml test_inputs/history_gen_yaml_error.json adaptor error
-test_inputs/existing_yaml_long.yaml test_inputs/history_gen_yaml_long.json
